feat: turn image & video renderer to typescript
