+++ src/PanoViewer/index.ts
@@ -2,6 +2,6 @@ import PanoViewer from "./PanoViewer";
 import { VERSION } from "../version";
 
 export {
-	PanoViewer,
-	VERSION
+  PanoViewer,
+  VERSION
 };
